Adds multiple filetypes support, accepting only image mime-type files, adds more  appropriate error messages and codes
